# Conflicts:
#	build.sbt
#	project/SparkVersionAxis.scala
#	spark-partition-sizing/src/test/scala/za/co/absa/spark/partition/sizing/DataFramePartitionerTest.scala
#	spark-partition-sizing/src/test/scala/za/co/absa/spark/partition/sizing/DummyDatasets.scala
#	spark-partition-sizing/src/test/scala/za/co/absa/spark/partition/sizing/sizer/FromDataframeSampleSizerTest.scala
#	spark-partition-sizing/src/test/scala/za/co/absa/spark/partition/sizing/sizer/FromDataframeSizerTest.scala
#	spark-partition-sizing/src/test/scala/za/co/absa/spark/partition/sizing/sizer/FromSchemaSizerTest.scala

+++ spark-partition-sizing/src/main/scala/za/co/absa/spark/partition/sizing/sizer/FromSchemaSizer.scala
@@ -26,7 +26,7 @@ import scala.util.control.TailCalls.{TailRec, done, tailcall}
 class FromSchemaSizer(implicit dataTypeSizes: DataTypeSizes) extends RecordSizer {
   private val zeroByteSize: ByteSize = 0
 
-  def performRowSizing(df: DataFrame, dfRecordCount: Option[Int] = None): ByteSize = {
+  def performRowSizing(df: DataFrame, dfRecordCount: Option[Long] = None): ByteSize = {
     val schema = df.schema
     structSize(schema, 1, done(zeroByteSize)).result
   }
